feat: v0.1.0 版本完成

+++ src/Rich.WebHook/Application/Users/IRichSession.cs
@@ -2,7 +2,7 @@ namespace Rich.WebHook.Application.Users;
 
 public interface IRichSession
 {
-    long? UserId { get;}
+    int? UserId { get;}
     string? UserName { get; }
     string? Email { get;  }
 }
